+++ src/common/js/global.js
@@ -2,4 +2,6 @@
  * pwix:startup-app-admin/src/common/js/global.js
  */
 
-pwixSAA = {};
+pwixSAA = {
+    _conf: {}
+};